Clean compilation after adding the assembly stage.

+++ clients/assembler/AssembledScalerEvent.h
@@ -56,7 +56,7 @@ class AssembledScalerEvent : public AssembledEvent
 private:
   unsigned long         m_startTime;  
   unsigned long         m_stopTime;
-  std::vector<unsigned long> m_scalers;
+  std::vector<uint32_t> m_scalers;
 
 public:
   AssembledScalerEvent(unsigned short node,
@@ -72,16 +72,16 @@ public:
   // Put/get scalers in the beast:
 
   void addScalers(void* pScalers, size_t count);
-  void addScalers(std::vector<unsigned long> scalers);
+  void addScalers(std::vector<uint32_t> scalers);
 
-  std::vector<unsigned long> getScalers() const;
-  unsigned long& operator[](unsigned int index);
+  std::vector<uint32_t> getScalers() const;
+  uint32_t& operator[](unsigned int index);
 
   // delegations to the m_scalers vector:
 
   size_t size() const;
-  std::vector<unsigned long>::iterator begin();
-  std::vector<unsigned long>::iterator end();
+  std::vector<uint32_t>::iterator begin();
+  std::vector<uint32_t>::iterator end();
  
 };
